requirements.txt
discord.py==1.7.3
Flask==2.1.2
flask_cors==3.0.10
mysql_connector_repackaged==0.3.1
python-dotenv==0.20.0
Werkzeug==2.1.2
